+++ pokemongo_bot/__init__.py
@@ -267,31 +267,6 @@ class PokemonGoBot(object):
         inventory_req = self.api.call()
         return inventory_req
 
-    def current_inventory(self):
-        inventory_req = self.player_service.get_inventory()
-        inventory_dict = inventory_req['responses']['GET_INVENTORY'][
-            'inventory_delta']['inventory_items']
-
-        user_web_inventory = 'web/inventory-%s.json' % (self.config.username)
-        with open(user_web_inventory, 'w') as outfile:
-            json.dump(inventory_dict, outfile)
-
-        # get player items stock
-        # ----------------------
-        items_stock = {x.value:0 for x in list(Item)}
-
-        for item in inventory_dict:
-            try:
-                # print(item['inventory_item_data']['item'])
-                item_id = item['inventory_item_data']['item']['item_id']
-                item_count = item['inventory_item_data']['item']['count']
-
-                if item_id in items_stock:
-                    items_stock[item_id] = item_count
-            except:
-                continue
-        return items_stock
-
     def item_inventory_count(self, id):
         inventory_req = self.player_service.get_inventory()
         inventory_dict = inventory_req['responses'][
